This file contain apex code.
In this file i have written a code witch is basically use to make callout to external website like here the external website is IMDB website. In the code i have created the object of HTTP class and HTTPREQUEST class with is basically use to perform outbound integration. And i have also created the object of Response class to get the response from IMDB website.